Update RT-TDDFT.md
I added some details to the formula that produces a Gaussian enveloped quasi-monochromatic pulse so that it is exact. Also, the parameter phase was missing. It's inclusion makes it straightforward to move from a delta-type kick field to a narrow Gaussian electric field, in order to avoid introducing nonphysical artifacts or instabilities in calculations of absorption spectra, as it is discussed in K. Lopata and N. Govind, J. Chem. Theory Comput. 2011, 7, 5, 1344–1355.

Former-commit-id: 7841a3d8321a1ef8e6bed4b778aa3eb162924fc4